Issue 16 - Move policy evaluation code from iSHARE-compliant authorization mode module to separate utility module (#43)
* Starting with util module for evaluating NGSI request policy parameters action, ID, type and attributes.
Functiond for action, ID and type implemented.
Open: attributes

* * Adding iSHARE module.
* Change policy validator code for usage of ngsi and ishare modules

* Removing paths entry
